+++ History.md
@@ -2,6 +2,7 @@ Unpublished
 ===========
 
   * example: measure distance with a HC-SR04 ultrasonic sensor
+  * initialize and terminate added to public API
 
 0.3.4 / May 15 2016
 ===================